backend/requirements.txt
# Core Django Framework
Django==4.2.24
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
django-allauth==65.11.2
django-cors-headers==4.9.0
django-ratelimit==4.1.0
django-timezone-field==7.1
django-celery-beat==2.8.1
django-filter==24.3

# Database & ORM
psycopg2-binary==2.9.10
sqlparse==0.5.3

# Task Queue & Background Processing
celery==5.5.3
redis==6.4.0
kombu==5.5.4
amqp==5.3.1
billiard==4.2.2
vine==5.1.0

# File Processing & Document Parsing
pypdf==6.1.1
PyPDF2==3.0.1
pdf2image==1.17.0
PyMuPDF==1.26.4
python-docx==1.2.0
openpyxl==3.1.5
xlsxwriter==3.2.9
et_xmlfile==2.0.0

# Data Processing & Analysis
pandas==2.3.2
numpy==2.2.6
lxml==6.0.1
beautifulsoup4==4.13.5

# Computer Vision & OCR
opencv-python==4.12.0.88
pytesseract==0.3.13
pillow==11.3.0

# AI & Machine Learning
openai==1.108.1
ollama==0.5.4
anthropic==0.40.0

# HTTP & API Clients
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
httplib2==0.31.0
urllib3==2.5.0

# Google APIs
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0

# Authentication & Security
pycryptodome==3.23.0
cryptography==46.0.1
PyJWT==2.10.1
oauthlib==3.3.1
requests-oauthlib==1.3.1

# Utilities & Helpers
python-decouple==3.8
python-dateutil==2.9.0.post0
python-crontab==3.3.0
cron_descriptor==2.0.6
tqdm==4.67.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52

# Communication & Messaging
python-telegram-bot==22.4

# External Integrations
splitwise==3.0.0

# Cloud Storage
boto3==1.35.50
botocore==1.35.50
s3transfer==0.10.3
django-storages==1.14.4

# Security & Encryption
cryptography>=41.0.0  # For encrypting stored PDF passwords

# Deployment & Production
gunicorn==23.0.0
whitenoise==6.11.0

# Development & Code Quality
ruff==0.13.1
vulture==2.3

# Core Dependencies
asgiref==3.9.1
packaging==25.0
pytz==2025.2
tzdata==2025.2
six==1.17.0
setuptools

# Low-level Dependencies
cffi==2.0.0
pycparser==2.23
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
distro==1.9.0
jiter==0.11.0
pyparsing==3.2.5
soupsieve==2.8
wcwidth==0.2.13
typing_extensions==4.15.0
typing-inspection==0.4.1
uritemplate==4.2.0
cachetools==5.5.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0
